requirements.txt
bcrypt==3.2.2
bidict==0.22.0
cffi==1.15.0
click==8.1.3
colorama==0.4.4
cycler==0.10.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-SocketIO==4.3.1
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.3.1
MarkupSafe==2.1.1
matplotlib==3.4.3
numpy==1.20.3
pandas==1.2.4
Pillow==8.3.1
pycparser==2.21
PyJWT==2.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2021.1
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
